Create separate implementations of classes for each supported framework
