+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/Views/Home/Index.cshtml
@@ -15,8 +15,6 @@
 
 <div class="grid-row">
     <div class="column-two-thirds">
-        
-        @Url.ProviderApprenticeshipServiceLink("");
         <h1 class="heading-xlarge">Add and update apprentice records as a training provider</h1>
 
         <p>This service is for training providers registered with the apprenticeship service to work with employers to:</p>
