+++ src/SFA.DAS.UrlHelper/SFA.DAS.ProviderUrlHelper.csproj
@@ -1,7 +1,6 @@
 ﻿ <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <!--<TargetFramework>netcoreapp2.2</TargetFramework>-->
     <TargetFrameworks>netstandard2.0;net462;netcoreapp2.2</TargetFrameworks>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
   </PropertyGroup>
@@ -20,7 +19,28 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Remove="Standard\**" />
+    <EmbeddedResource Remove="Standard\**" />
+    <None Remove="Standard\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="SFA.DAS.AutoConfiguration" Version="2.0.1" />
+  </ItemGroup>
+
+   <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp2.2'">
+     <Folder Include="Core\" />
+     <Compile Remove="Framework\*" />
+   </ItemGroup>
+
+   <ItemGroup Condition="'$(TargetFramework)' == 'net462'">
+     <Folder Include="Framework\" />
+     <Compile Remove="Core\*" />
+   </ItemGroup>
+
+   <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+     <Compile Remove="Core\*" />
+     <Compile Remove="Framework\*" />
   </ItemGroup>
 
 </Project>
